+++ modules/resources/components/submit-resource-welcome-card.tsx
@@ -1,4 +1,4 @@
-import { auth, currentUser } from '@clerk/nextjs';
+import { currentUser } from '@clerk/nextjs';
 
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { cn } from '@/lib/utils';
